BUG/MEDIUM: srv-state: always ensure there's a warmup task before manipulating it
Raghu Udiyar reported an issue by which haproxy segfaults on startup
when reloading with a server that was disabled + wears a slowstart in
the config and which is enabled in the state file. What happens is that
when reading the state file, the server seen as up is started and the
warmup task is woken up. But this is performed before this task is
allocated because the server was marked disabled.

There's some of a chicken-an-egg issue to deal with here. The problem
doesn't exist in 1.8 and above since the initialization sequence
significantly changed, and these changes are unsuitable for 1.7.

The most reasonable solution for 1.7 is to protect the few places where
there is an attempt to wake the warmup task up by first checking if this
task exists. When loading the state from a file, it will be ignored, but
then upon initialization completion, when the task is allocated, it will
be automatically woken up.

A backport to 1.6 may be needed though this hasn't been checked yet.