BUG/MINOR: poll: too large size allocation for FD events
Commit 80da05a ("MEDIUM: poll: do not use FD_* macros anymore") which
appeared in 1.5-dev18 and which was backported to 1.4.23 made explicit
use of arrays of FDs mapped to unsigned ints. The problem lies in the
allocated size for poll(), as the resulting size is in bits and not
bytes, resulting in poll() arrays being 8 times larger than necessary!

In practice poll() is not used on highly loaded systems, explaining why
nobody noticed. But it definetely has to be addressed.

This fix needs to be backported to all stable versions.

(cherry picked from commit cc35923c329bbadc78c1c026ce4e45e115852abe)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 13165eed6abb8569d56f6eec404f4462d0cd7115)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>