MINOR: servers: Don't report duplicate dyncookies for disabled servers.
Especially with server-templates, it can happen servers starts with a
placeholder IP, in the disabled state. In this case, we don't want to report
that the same cookie was generated for multiple servers. So defer the test
until the server is enabled.

This should be backported to 1.8.

(cherry picked from commit e9bad0a9361a761f2f304dcca8ddd997c05627c2)
Signed-off-by: Willy Tarreau <w@1wt.eu>