BUG/MEDIUM: checks: Don't try to release undefined conn_stream when a check is freed
When a healt-check is released, the attached conn_stream may be undefined. For
instance, this happens when 'no-check' option is used on a server line. So we
must check it is defined before trying to release it.

This patch must be backported in 1.8.

(cherry picked from commit 23d86d157edd6da8d3fe40d7883a3d2beeed3189)
Signed-off-by: Willy Tarreau <w@1wt.eu>