BUG/MINOR: threads: Update labels array because of changes in lock_label enum
Recent changes to the enum were not synchronized with the lock debugging
code. Now we use a switch/case instead of an array so that the compiler
throws a warning if there is any inconsistency.

To be backported to 1.8 (at least to add the START entry).

(cherry picked from commit f51bac2ba83a27a75ef6bab58bfde6489b999c49)
Signed-off-by: Willy Tarreau <w@1wt.eu>