BUG/MINOR: ssl/threads: Make management of the TLS ticket keys files thread-safe
A TLS ticket keys file can be updated on the CLI and used in same time. So we
need to protect it to be sure all accesses are thread-safe. Because updates are
infrequent, a R/W lock has been used.

This patch must be backported in 1.8

(cherry picked from commit 16f45c87d5f9bf57fa4e7e71546bce352a727425)
Signed-off-by: Willy Tarreau <w@1wt.eu>