BUG/MEDIUM: buffer: Fix the wrapping case in bi_putblk
When the block of data need to be split to support the wrapping, the start of
the second block of data was wrong. We must be sure to skup data copied during
the first memcpy.

This patch must be backported to 1.8.

(cherry picked from commit ca6ef506610e9d78f99b7ab2095ce0f8a47e18df)
Signed-off-by: Willy Tarreau <w@1wt.eu>