BUG/MEDIUM: spoe: Remove idle applets from idle list when HAProxy is stopping
In the SPOE applet's handler, when an applet is switched from the state IDLE to
PROCESSING, it is removed for the list of idle applets. But when HAProxy is
stopping, this applet can be switched to DISCONNECT. In this case, we also need
to remove it from the list of idle applets. Else the applet is removed but still
present in the list. It could lead to a segmentation fault or an infinite loop,
depending the code path.

(cherry picked from commit 7d9f1ba246055046eed547fa35aa546683021dce)
[wt: adapted context for 1.8]
Signed-off-by: Willy Tarreau <w@1wt.eu>