BUG/MINOR: lua: the function returns anything
The output of these function indicates that one element is pushed in
the stack, but no element is set in the stack. Actually, if anyone
read the value returned by this function, is gets "something"
present in the stack.

This patch is a complement of these one: 119a5f10e47f3507e58116

The LuaSocket documentation tell anything about the returned value,
but the effective code set an integer of value one.

   https://github.com/diegonehab/luasocket/blob/316a9455b9cb4637fe6e62b20fbe05f5141fec54/src/timeout.c#L172

Thanks to Tim for the bug report.

This patch should be backported in all version from 1.6

(cherry picked from commit e9636f192a5c312424b7598ca182b0425827b08b)
Signed-off-by: Willy Tarreau <w@1wt.eu>