BUG/MINOR: force-persist and ignore-persist only apply to backends
>From the very first day of force-persist and ignore-persist features,
they only applied to backends, except that the documentation stated it
could also be applied to frontends.

In order to make it clear, the documentation is updated and the parser
will raise a warning if the keywords are used in a frontend section.

This patch should be backported up to the 1.5 branch.

(cherry picked from commit 4288c5a9d81f89f65c02574a434b02e98eea943c)
Signed-off-by: Willy Tarreau <w@1wt.eu>