BUG/MINOR: spoe: Don't release the context buffer in .check_timeouts callbaclk
It is better to let spoe_stop_processing release this buffer because, in
.check_timeouts callback, we lack information to know if it should be release or
not. For instance, if the processing timeout is reached while the SPOE applet
receives the reply, it is preferable to ignore the timeout and process the
result.

This patch should be backported in 1.8.

(cherry picked from commit ac580608d712e1e8c351d84dfda6e5ac773812df)
Signed-off-by: Willy Tarreau <w@1wt.eu>