BUG/MINOR: email-alert: Set the mailer port during alert initialization
Since the commit 2f3a56b4f ("BUG/MINOR: tcp-check: use the server's service port
as a fallback"), email alerts stopped working because the mailer's port was
overriden by the server's port. Remember, email alerts are defined as checks
with specific tcp-check rules and triggered on demand to send alerts. So to send
an email, a check is executed. Because no specific port's was defined, the
server's one was used.

To fix the bug, the ports used for checks attached an email alert are explicitly
set using the mailer's port. So this port will be used instead of the server's
one.

In this patch, the assignement to a default port (587) when an email alert is
defined has been removed. Indeed, when a mailer is defined, the port must be
defined. So the default port was never used.

This patch must be backported in 1.8.

(cherry picked from commit b797ae1f15b0fb97d1ae8550d37c3fd58a24f5b3)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 12db95e9962b2e2902a341e49754f27f20fe3ace)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>